Added Kingdom of Yellowstone
added 1 kingdom, 4 duchies, and 13 new provinces covering the area of
western Wyoming, north eastern Utah and south eastern Idaho. Added a new
culture of frontiersman. I was going to name it Mountain Man, but
thought it might be confused with Mountainer. Most of the provinces
added are gaian religion, but Idaho Falls and Pocatello area are Mormon,
and Twin Falls area is Cetic.

This is my first addition to this project and so you may need to change
some of my additions. I left room so that a kingdom can be added between
Yellowstone and Lakotah, and another kingdom between Yellowstone and
Jefferson.