Slight buff for the Americanists in San Antonio
The implication of the Airmen is they're the last of a dying breed,
descendents of Pilots at the Air base at San Antonio. Also they tend to
rekt immediately in the game. This is them getting thrown a bone, 2 of
their 4 counties starting as Americanist rather than Cult of Saints.