Added new character in Chapel Hill
Chapel Hill is one of the Evangelical Pentarchies, so it makes sense to give it a defined holder at the start of the game.

This new character is pretty much just the Reverend Billy Graham if he was born in 2618 rather than 1918. He was the first person that came to mind when trying to think of a notable Protestant religious figure from North Carolina. Adding someone who's pretty much a direct copy of a historical figure isn't unprecedented in AtE - see the Count of Bismark.