Added events that let Occultist characters choose lore traits on adulthood
This is how these events should work:
occult.1 only fires after reformation for characters who already follow a school of lore. This event lets people who follow an unreformed school choose which of the reformed schools to follow.

occult.2 fires on adulthood if you're an unreformed Occultist. It lets you choose which school of Occultist lore you wish to follow for the rest of your life.

occult.3 functions similarly to the above, but for reformed Occultists.

I sincerely hope this event doesn't get screwed by the America Restoration event, which has already proven to be cursed.