West Indies Redux
The flag of the West Indian Federation was already in the game, as the flag of the Kingdom of Antilles. However, now that there is an actual kingdom named the Kingdom of the West Indies, it seems right to change the flag accordingly. k_antilles now uses the old flag of the Netherlands Antilles.

Also, fixed Colima.