changed barony set-up in Beverly Hills
Justification:
First, some background on how LA works. The city of Los Angeles and LA County are two completely different entities. Cities like Long Beach, Pasadena, Compton, and Beverly Hills are just that - separate cities entirely. They just all happen to be in the same county, which is also named Los Angeles.

Now that we have that out of the way, here's my justification for this change. Every single other barony holding in Los Angeles represents a different city in LA County. Long Beach, Santa Monica, Pasadena, Compton, Inglewood and Burbank are all independent cities in LA county.

The only exception to this rule is the barony of Hollywood. Hollywood isn't a separate city - it's just a particularly notable neighborhood in the city of Los Angeles. It's true that it *used* to be its own city, but it got gobbled up by LA well before World War 2 and subsequent independence movements never built up any steam. Basically, besides the Hollywood meme, there's no reason to include it as an entire separate barony.

Now, for Beverly Hills. Although it's relatively small in size, it punches way above its weight. It's easily the richest and most famous city in LA county that isn't Los Angeles itself. Even if we completely ignore its pop culture impact, the fact that it's basically where all of the rich and famous decide to live makes it the perfect candidate for being the capital holding of a merchant republic - moreso than Hollywood itself, which is actually below-average in terms of median income. Although it's where all of the famous landmarks are located and where a bunch of film studios are based, it's not where most of these people actually live. And given that the mansions of merchant families are portrayed as separate holdings, I'd say that where people live is a fairly important detail to include. (All of this was true even before the modern-day - Beverly Hills has had this reputation since the 50s.)

Also I'm from Beverly Hills and I'm salty that it's not in the mod.