moved Garinagu culture to the Arawak culture group
Garinagu is literally the name for the Garifuna people in the Garifuna language, and the Garifuna language is an Arawak language, so why the hell is it in the Caribbanglo group? The Arawak group only has a single member at the moment as well. Therefore, I feel justified in doing this from both a flavor and gameplay standpoint.