Changed event images for most Brazil events, integrated the new statuses and hl_silk_route_events
The new statuses are functionally similar to the normal Civil War status, but they will always introduce a new emperor of the relevant religion and do not have a chance of restoring a pretender.

maybe I should change the raiders in hl_sil_route_events to pirates, given the coastal nature of the caribbean trade route, but this should do for now.