More compatibility work
Also fixed missing Brazil event file
TODO: finish rest of files modified on 12/6. stopped at on_action because dear god that's gonna be as much work as all those other files put together.

then I have to move on to all the files changed in November...

and this is just for the events folder...