Added Scooby-Doo flavor event
This event chain fires for Norville of Rutland, and involves you teaming up with a band of teenage investigators to uncover the truth behind a haunting in a nearby village.

Also:
*New artifact added: Ancient Rosary
*God names changed for various religions
*Consumerists use crusades instead of prepared invasions
*Removed errant references to Reformed Rust Cultist
*Adjusted flags of several titles
*Adjusted titles for Consumerist Merchant Republics
*Adjusted modifiers for picking SH traits
*Adjusted some additional localization

if you're wondering why I used the Norwegian CoA for Bella Coola, apparently the area has a long history of Norwegian settlement. it's better than a company logo at any rate