New Baronies, Cultural Placenames in Northwest
Added a collection of barony names and cultural placenames from Haida, Shuswap and Canuck cultures to the counties of the reworked areas in British Columbia and Northern Alberta. The only two affected files are cascadia.txt and thunderland.txt. No localization has been done for these areas yet.

Additionally, moved county of Camano from the Duchy of Vancouver to the Duchy of Puget, to reflect the natural barrier between Bellingham Bay and Camano produced by the Chuckanut Mountains.