Added religious branch traits for Shinto
based on the suggestions made in the same reddit post I got the hindu stuff from. bonuses are based on their theological origins - i.e., koshitsu gives vassal opinion bc it's focused on worship of the emperor, ryobu gives learning because it's based on syncretism with vajrayana buddhism, minzoku gives prestige because it's based on the more popular forms of shinto worship, and shugendo gets piety because it's based on mountain monk mysticism.

i also changed the descriptions for the hindu branch traits so they work for random world. this, incidentally, is why none of the shinto traits go into detail.