Updated d_nestorian
Updated the regal names for d_nestorian to be Assyrian spellings when if possible, if not then Hebrew, and if not that than Greek.  The reason for the later two are Assyrians seems to borrow names from them (and translate them into their language).

The numbers beside the names are reflective of the amount of how many of them had appeared over the years as leader of the church, like how there has been 7 Yoseps so far in real life. The numbers affect how often they appear in-game as head of d_nestorian.