Updated Utahn names
Meriador in the community server pointed out that a number of utahn names were either not fitting or in very poor taste for Mormons (the equivalent of naming a kid Lucifer or Judas) and provided a list of replacements.
Also there's a single name that's two words and hopefully should work with an nonbeaking space but I might have broken that.