Replaced religion_group = pirate with religion_group = nautical
Honestly, I'm surprised this mod kept working so long with a religion group that had the exact same name as a trait. We really really really had to separate out those two, because that has the potential to cause all sorts of problems under the hood. I think I got every reference to the pirate religious group, but I might be wrong. I also fixed a few other instances where the religion group was accidentally treated as a standalone religion (another reason to change it!)